Robots/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.9
djangorestframework==3.14.0
kombu==5.3.4
progressbar2==4.3.2
prompt-toolkit==3.0.43
psycopg==3.1.16
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
python-utils==3.8.1
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
XlsxWriter==3.1.9
oauthlib==2.0.2
django-oauth-toolkit==1.0.0